+++ core/gethwrappers/generation/generate_link/util/json_normalization.go
@@ -1,4 +1,4 @@
-package json_utils
+package util
 
 import (
 	"bufio"
